Consolidate _clamp_score into common.py
Move _clamp_score() from evaluate_model.py, score_dataset.py, and
generate_distillation_data.py into common.py. All three already import
from common, so this eliminates the triplication.
