+++ plugin/skills/microsoft-foundry/finetuning/scripts/score_dataset.py
@@ -35,7 +35,7 @@ except (AttributeError, OSError):
 import time
 from concurrent.futures import ThreadPoolExecutor, as_completed
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-from common import HelpOnErrorParser, get_clients
+from common import HelpOnErrorParser, get_clients, _clamp_score
 
 
 QUALITY_PROMPT = """You are a data quality assessor for machine learning training data.
@@ -63,21 +63,6 @@ DEFAULT_DIMENSIONS = {
     "relevance": "Does the output directly address the user's request?",
     "quality": "Is the output well-written, well-formatted, and professional?",
 }
-
-
-def _clamp_score(v, default=0):
-    """Clamp a judge score to [1, 10]. Returns `default` for missing/non-numeric values.
-
-    LLM judges occasionally return out-of-range integers (e.g., 15) or non-numeric
-    strings ("high"). 0 is the sentinel for "missing/failed" so callers can filter
-    via `score > 0`.
-    """
-    if v is None:
-        return default
-    try:
-        return max(1, min(10, int(v)))
-    except (ValueError, TypeError):
-        return default
 
 
 def score_example(client, model, user_content, assistant_content, dimensions):
